Keep spawn options helper internal
